+++ cypress/integration/professions/search/search.spec.ts
@@ -21,9 +21,13 @@ describe('Searching a profession', () => {
   });
 
   it('Professions are sorted alphabetically', () => {
-    cy.get('h2').then((elements) => {
+    cy.get('a.rpr-listing__profession-title').then((elements) => {
       const names = elements.map((_, element) => element.innerText).toArray();
-      cy.wrap(names).should('deep.equal', names.sort());
+
+      cy.wrap(names).should(
+        'deep.equal',
+        [...names].sort((a: string, b: string) => a.localeCompare(b)),
+      );
     });
   });