+++ src/professions/profession-versions.service.ts
@@ -164,12 +164,11 @@ export class ProfessionVersionsService {
   }
 
   async searchLive(filter: FilterInput): Promise<Profession[]> {
-    let query = this.versionsWithJoins().where(
-      'professionVersion.status = :status',
-      {
+    let query = this.versionsWithJoins()
+      .orderBy('profession.name')
+      .where('professionVersion.status = :status', {
         status: ProfessionVersionStatus.Live,
-      },
-    );
+      });
 
     if (filter.keywords?.length) {
       const ids = await this.searchService.search(filter.keywords);
@@ -213,7 +212,7 @@ export class ProfessionVersionsService {
 
   async allWithLatestVersion(): Promise<Profession[]> {
     const versions = await this.versionsWithJoins()
-      .distinctOn(['professionVersion.profession'])
+      .distinctOn(['professionVersion.profession', 'profession.name'])
       .where('professionVersion.status IN(:...status)', {
         status: [
           ProfessionVersionStatus.Live,
@@ -222,7 +221,7 @@ export class ProfessionVersionsService {
         ],
       })
       .orderBy(
-        'professionVersion.profession, professionVersion.created_at',
+        'profession.name, professionVersion.profession, professionVersion.created_at',
         'DESC',
       )
       .getMany();
@@ -236,7 +235,7 @@ export class ProfessionVersionsService {
     organisation: Organisation,
   ): Promise<Profession[]> {
     const versions = await this.versionsWithJoins()
-      .distinctOn(['professionVersion.profession'])
+      .distinctOn(['professionVersion.profession', 'profession.name'])
       .where('professionVersion.status IN(:...status)', {
         status: [
           ProfessionVersionStatus.Live,
@@ -248,7 +247,7 @@ export class ProfessionVersionsService {
         organisationId: organisation.id,
       })
       .orderBy(
-        'professionVersion.profession, professionVersion.created_at',
+        'profession.name, professionVersion.profession, professionVersion.created_at',
         'DESC',
       )
       .getMany();
@@ -260,13 +259,13 @@ export class ProfessionVersionsService {
 
   async latestVersion(profession: Profession): Promise<ProfessionVersion> {
     return await this.versionsWithJoins()
-      .distinctOn(['professionVersion.profession'])
+      .distinctOn(['professionVersion.profession', 'profession.name'])
       .where('professionVersion.status IN(:...status)', {
         status: [ProfessionVersionStatus.Live, ProfessionVersionStatus.Draft],
       })
       .where({ profession: profession })
       .orderBy(
-        'professionVersion.profession, professionVersion.created_at',
+        'profession.name, professionVersion.profession, professionVersion.created_at',
         'DESC',
       )
       .getOne();